Refactor authenticators extract Docker and containerd to individual files
Context
---

We currently suport 2 authentication sources. Docker and containerD.

Problems
---

1. Kubelet has a third one: [CredentialProviderConfig]

To be consistent, noe needs to implement all of those to be able to
lookup the values in the relevant place.

2. The current containerD authentication is mixed with the Docker one (file
and image pull secret ones).

Goal
---

Get clearer isolation between the different authentication mechanisms
Ease implementing new authentication methods

[CredentialProviderConfig]: https://kubernetes.io/docs/tasks/administer-cluster/kubelet-credential-provider/
